member 도메인 리팩토링

+++ src/main/java/cloneproject/Instagram/domain/member/service/MemberAuthService.java
@@ -1,5 +1,7 @@
 package cloneproject.Instagram.domain.member.service;
 
+import static cloneproject.Instagram.global.error.ErrorCode.*;
+
 import java.util.List;
 import java.util.Optional;
 
@@ -12,17 +14,17 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import cloneproject.Instagram.domain.member.dto.JwtDto;
+import cloneproject.Instagram.domain.member.dto.LoginDevicesDTO;
 import cloneproject.Instagram.domain.member.dto.LoginRequest;
 import cloneproject.Instagram.domain.member.dto.LoginWithCodeRequest;
-import cloneproject.Instagram.domain.member.dto.LoginedDevicesDTO;
 import cloneproject.Instagram.domain.member.dto.RegisterRequest;
 import cloneproject.Instagram.domain.member.dto.ResetPasswordRequest;
 import cloneproject.Instagram.domain.member.dto.UpdatePasswordRequest;
 import cloneproject.Instagram.domain.member.entity.Member;
 import cloneproject.Instagram.domain.member.entity.redis.RefreshToken;
 import cloneproject.Instagram.domain.member.exception.AccountMismatchException;
-import cloneproject.Instagram.domain.member.exception.PasswordResetFailException;
 import cloneproject.Instagram.domain.member.exception.InvalidJwtException;
+import cloneproject.Instagram.domain.member.exception.PasswordResetFailException;
 import cloneproject.Instagram.domain.member.repository.MemberRepository;
 import cloneproject.Instagram.domain.search.entity.SearchMember;
 import cloneproject.Instagram.domain.search.repository.SearchMemberRepository;
@@ -30,13 +32,11 @@ import cloneproject.Instagram.global.error.exception.EntityAlreadyExistException
 import cloneproject.Instagram.global.error.exception.EntityNotFoundException;
 import cloneproject.Instagram.global.util.AuthUtil;
 import cloneproject.Instagram.global.util.JwtUtil;
-import cloneproject.Instagram.infra.geoip.dto.GeoIP;
 import cloneproject.Instagram.infra.geoip.GeoIPLocationService;
+import cloneproject.Instagram.infra.geoip.dto.GeoIP;
 import io.jsonwebtoken.JwtException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-
-import static cloneproject.Instagram.global.error.ErrorCode.*;
 
 @Slf4j
 @Service
@@ -55,10 +55,7 @@ public class MemberAuthService {
 
 	@Transactional(readOnly = true)
 	public boolean checkUsername(String username) {
-		if (memberRepository.existsByUsername(username)) {
-			return false;
-		}
-		return true;
+		return !memberRepository.existsByUsername(username);
 	}
 
 	@Transactional
@@ -66,18 +63,18 @@ public class MemberAuthService {
 		if (memberRepository.existsByUsername(registerRequest.getUsername())) {
 			throw new EntityAlreadyExistException(USERNAME_ALREADY_EXIST);
 		}
-		String username = registerRequest.getUsername();
+		final String username = registerRequest.getUsername();
 
 		if (!emailCodeService.checkEmailCode(username, registerRequest.getEmail(), registerRequest.getCode())) {
 			return false;
 		}
 
-		Member member = registerRequest.convert();
-		String encryptedPassword = bCryptPasswordEncoder.encode(member.getPassword());
+		final Member member = registerRequest.convert();
+		final String encryptedPassword = bCryptPasswordEncoder.encode(member.getPassword());
 		member.setEncryptedPassword(encryptedPassword);
 		memberRepository.save(member);
 
-		SearchMember searchMember = new SearchMember(member);
+		final SearchMember searchMember = new SearchMember(member);
 		searchMemberRepository.save(searchMember);
 
 		return true;
@@ -93,15 +90,16 @@ public class MemberAuthService {
 	@Transactional
 	public JwtDto login(LoginRequest loginRequest, String device, String ip) {
 		try {
-			UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(
-					loginRequest.getUsername(), loginRequest.getPassword());
-			Authentication authentication = authenticationManagerBuilder.getObject().authenticate(authenticationToken);
-			JwtDto jwtDto = jwtUtil.generateTokenDto(authentication);
+			final UsernamePasswordAuthenticationToken authenticationToken = new UsernamePasswordAuthenticationToken(
+				loginRequest.getUsername(), loginRequest.getPassword());
+			final Authentication authentication = authenticationManagerBuilder.getObject()
+				.authenticate(authenticationToken);
+			final JwtDto jwtDto = jwtUtil.generateTokenDto(authentication);
 
-			GeoIP geoIP = geoIPLocationService.getLocation(ip);
+			final GeoIP geoIP = geoIPLocationService.getLocation(ip);
 
 			refreshTokenService.addRefreshToken(Long.valueOf(authentication.getName()), jwtDto.getRefreshToken(),
-					device, geoIP);
+				device, geoIP);
 			return jwtDto;
 		} catch (BadCredentialsException e) {
 			throw new AccountMismatchException();
@@ -109,7 +107,7 @@ public class MemberAuthService {
 	}
 
 	@Transactional
-	public Optional<JwtDto> reisuue(String refreshTokenString) {
+	public Optional<JwtDto> reissue(String refreshTokenString) {
 		if (!jwtUtil.validateRefeshJwt(refreshTokenString)) {
 			throw new InvalidJwtException();
 		}
@@ -119,26 +117,28 @@ public class MemberAuthService {
 		} catch (JwtException e) {
 			throw new InvalidJwtException();
 		}
+		final Member member = authUtil.getLoginMember();
 
-		Optional<RefreshToken> refreshToken = refreshTokenService
-				.findRefreshToken(Long.valueOf(authentication.getName()), refreshTokenString);
+		final Optional<RefreshToken> refreshToken = refreshTokenService.findRefreshToken(member.getId(),
+			refreshTokenString);
 		if (refreshToken.isEmpty()) {
 			return Optional.empty();
 		}
-		JwtDto jwtDto = jwtUtil.generateTokenDto(authentication);
+		final JwtDto jwtDto = jwtUtil.generateTokenDto(authentication);
 		refreshTokenService.updateRefreshToken(refreshToken.get(), jwtDto.getRefreshToken());
 		return Optional.of(jwtDto);
 	}
 
 	@Transactional
 	public void updatePassword(UpdatePasswordRequest updatePasswordRequest) {
-		Member member = authUtil.getLoginMember();
-		boolean oldPasswordCorrect = bCryptPasswordEncoder.matches(updatePasswordRequest.getOldPassword(),
-				member.getPassword());
-		if (!oldPasswordCorrect) {
+		final Member member = authUtil.getLoginMember();
+		if (!bCryptPasswordEncoder.matches(updatePasswordRequest.getOldPassword(), member.getPassword())) {
 			throw new AccountMismatchException();
 		}
-		String encryptedPassword = bCryptPasswordEncoder.encode(updatePasswordRequest.getNewPassword());
+		if (updatePasswordRequest.getOldPassword().equals(updatePasswordRequest.getNewPassword())) {
+			throw new EntityAlreadyExistException(PASSWORD_ALREADY_EXIST);
+		}
+		final String encryptedPassword = bCryptPasswordEncoder.encode(updatePasswordRequest.getNewPassword());
 		member.setEncryptedPassword(encryptedPassword);
 		memberRepository.save(member);
 	}
@@ -150,33 +150,38 @@ public class MemberAuthService {
 
 	@Transactional
 	public JwtDto resetPassword(ResetPasswordRequest resetPasswordRequest, String device, String ip) {
-		Member member = memberRepository.findByUsername(resetPasswordRequest.getUsername())
-				.orElseThrow(() -> new EntityNotFoundException(MEMBER_NOT_FOUND));
+		final Member member = memberRepository.findByUsername(resetPasswordRequest.getUsername())
+			.orElseThrow(() -> new EntityNotFoundException(MEMBER_NOT_FOUND));
+
 		if (!emailCodeService.checkResetPasswordCode(resetPasswordRequest.getUsername(),
-				resetPasswordRequest.getCode())) {
+			resetPasswordRequest.getCode())) {
 			throw new PasswordResetFailException();
 		}
-		String encryptedPassword = bCryptPasswordEncoder.encode(resetPasswordRequest.getNewPassword());
+		if (bCryptPasswordEncoder.matches(resetPasswordRequest.getNewPassword(), member.getPassword())) {
+			throw new EntityAlreadyExistException(PASSWORD_ALREADY_EXIST);
+		}
+
+		final String encryptedPassword = bCryptPasswordEncoder.encode(resetPasswordRequest.getNewPassword());
 		member.setEncryptedPassword(encryptedPassword);
 		memberRepository.save(member);
-		JwtDto jwtDto = login(
-				new LoginRequest(resetPasswordRequest.getUsername(), resetPasswordRequest.getNewPassword()), device,
-				ip);
+		final JwtDto jwtDto = login(
+			new LoginRequest(resetPasswordRequest.getUsername(), resetPasswordRequest.getNewPassword()), device,
+			ip);
 		emailCodeService.deleteResetPasswordCode(resetPasswordRequest.getUsername());
 		return jwtDto;
 	}
 
 	@Transactional
 	public JwtDto loginWithCode(LoginWithCodeRequest loginRequest, String device, String ip) {
-		Member member = memberRepository.findByUsername(loginRequest.getUsername())
-				.orElseThrow(() -> new EntityNotFoundException(MEMBER_NOT_FOUND));
+		final Member member = memberRepository.findByUsername(loginRequest.getUsername())
+			.orElseThrow(() -> new EntityNotFoundException(MEMBER_NOT_FOUND));
 		if (!emailCodeService.checkResetPasswordCode(loginRequest.getUsername(), loginRequest.getCode())) {
 			throw new PasswordResetFailException();
 		}
-		JwtDto jwtDto = jwtUtil.generateTokenDto(jwtUtil.getAuthenticationWithMember(member.getId().toString()));
+		final JwtDto jwtDto = jwtUtil.generateTokenDto(jwtUtil.getAuthenticationWithMember(member.getId().toString()));
 		emailCodeService.deleteResetPasswordCode(loginRequest.getUsername());
 
-		GeoIP geoIP = geoIPLocationService.getLocation(ip);
+		final GeoIP geoIP = geoIPLocationService.getLocation(ip);
 
 		refreshTokenService.addRefreshToken(member.getId(), jwtDto.getRefreshToken(), device, geoIP);
 		return jwtDto;
@@ -191,18 +196,19 @@ public class MemberAuthService {
 	public void logout(String refreshToken) {
 		try {
 			refreshTokenService.deleteRefreshTokenWithValue(authUtil.getLoginMemberId(), refreshToken);
-		} catch (InvalidJwtException e) {
-			return;
+		} catch (InvalidJwtException ignored) {
+
 		}
 	}
 
-	public List<LoginedDevicesDTO> getLoginedDevices() {
-		Member member = authUtil.getLoginMember();
-		return refreshTokenService.getLoginedDevices(member.getId());
+	public List<LoginDevicesDTO> getLoginDevices() {
+		final Member member = authUtil.getLoginMember();
+		return refreshTokenService.getLoginDevices(member.getId());
 	}
 
 	@Transactional
 	public void logoutDevice(String tokenId) {
 		refreshTokenService.deleteRefreshTokenWithId(authUtil.getLoginMemberId(), tokenId);
 	}
+
 }
